refactor: (#26) CommentSection의 중복된 훅 제거 및 컴포넌트로부터 훅 분리
